backend/requirements.txt
asgiref==3.8.1
boto3
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
gevent
packaging==24.1
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
